+++ src/screens/Home.tsx
@@ -1,0 +1,21 @@
+import { type FC } from "react";
+import { StyleSheet, Text, View } from "react-native";
+import { StatusBar } from "expo-status-bar";
+
+export const HomeScreen: FC = () => {
+  return (
+    <View style={styles.container}>
+      <Text>Open up App.tsx to start working on your app! home</Text>
+      <StatusBar style="auto" />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+});
